requirements.txt
fastapi
uvicorn
markdown
jinja2
deta
python-dotenv
aiofiles
python-multipart
lxml
python-frontmatter==1.0.0